remove kubeseal cli depedency
