+++ main.go
@@ -69,7 +69,6 @@ func main() {
 }
 
 func setupRouter(coreClient corev1.CoreV1Interface, ssClient ssClient.BitnamiV1alpha1Interface, cfg *config.Config) *gin.Engine {
-	// sealer := seal.New(cfg.KubesealArgs)
 	sealer, err := seal.NewAPISealer("sealed-secrets", "sealed-secrets", "")
 	if err != nil {
 		log.Fatalf("Setup sealer: %s", err.Error())
